Corrige mais 2 bugs de descompilação: rarity não-inicializada e literais byte
BuriedTreasureService.tryFind(): 'rarity' só era atribuída dentro do
ramo mais profundo do ternário aninhado (CFR perdeu as outras
atribuições) — reescreve como cadeia de ternários linear e correta.

set(key, PersistentDataType.BYTE, 1): literal 'int' não infere para
'Byte' em posição de argumento genérico (diferente de contexto de
atribuição direta) — usa (byte) 1 explicitamente, como já fazia
CombatAbilityService.

+++ src/main/java/dev/icaro/foodtooltips/mining/GemService.java
@@ -137,7 +137,7 @@ implements Listener {
             }
         }
         c.getPersistentDataContainer().set(this.dataKey, PersistentDataType.STRING, String.join((CharSequence)";", records));
-        c.getPersistentDataContainer().set(this.generationKey, PersistentDataType.BYTE, 1);
+        c.getPersistentDataContainer().set(this.generationKey, PersistentDataType.BYTE, (byte) 1);
     }
 
     @EventHandler(priority=EventPriority.HIGHEST, ignoreCancelled=true)
